+++ src/physics/MeshBody.cpp
@@ -41,13 +41,13 @@ MeshBody::MeshBody(Transformation const & transformation, Vector3 const * veloci
 	auto const & indices = mesh.GetIndices();
 	
 	dGeomTriMeshDataBuildSingle(_mesh_data,
-		vertices.front().pos.GetAxes(), sizeof(Mesh::value_type), vertices.size(),
-		indices.data(), indices.size(), sizeof(Mesh::index_type));
+		vertices.front().pos.GetAxes(), static_cast<int>(sizeof(Mesh::value_type)), static_cast<int>(vertices.size()),
+		indices.data(), static_cast<int>(indices.size()), static_cast<int>(sizeof(Mesh::index_type)));
 
 	dGeomTriMeshSetData(_collision_handle, _mesh_data);
 	
 #if defined(CRAG_DEBUG)
-	_num_triangles = indices.size();
+	_num_triangles = static_cast<int>(indices.size());
 #endif
 
 	// set mass
